+++ .gitignore
@@ -20,5 +20,5 @@ erl_crash.dump
 *.ez
 
 # Ignore package tarball (built via "mix hex.build").
-hex_crypto-*.tar
+pkcs5-*.tar
