+++ README.md
@@ -1,3 +1,5 @@
-# HexCrypto
+# PKCS5
 
-Crypto libraries used by Hex
+PKCS #5: Password-Based Cryptography Specification Version 2.0
+
+  See: https://tools.ietf.org/html/rfc2898
