+++ lib/pkcs5/public_key.ex
@@ -1,4 +1,4 @@
-defmodule HexCrypto.PublicKey do
+defmodule PKCS5.PublicKey do
   @doc """
   Decodes a public key and raises if the key is invalid.
   """
