+++ lib/pkcs5/utils.ex
@@ -1,4 +1,4 @@
-defmodule HexCrypto.Utils do
+defmodule PKCS5.Utils do
   def base64url_encode(binary) do
     try do
       Base.url_encode64(binary, padding: false)
