+++ mix.exs
@@ -1,9 +1,9 @@
-defmodule HexCrypto.MixProject do
+defmodule PKCS5.MixProject do
   use Mix.Project
 
   def project do
     [
-      app: :hex_crypto,
+      app: :pkcs5,
       version: "0.1.0",
       elixir: "~> 1.6",
       start_permanent: Mix.env() == :prod,
