--- test/hex_crypto_test.exs
@@ -1,3 +1,0 @@
-defmodule HexCryptoTest do
-  use ExUnit.Case, async: true
-end
