+++ test/pkcs5_test.exs
@@ -1,6 +1,6 @@
-defmodule HexCrypto.PKCS5Test do
+defmodule PKCS5Test do
   use ExUnit.Case, async: true
-  import HexCrypto.PKCS5
+  import PKCS5
 
   defp check_vectors(list) do
     Enum.each(list, fn {password, salt, iterations, derived_key_length, hash, derived_key} ->
